requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
decorator==5.1.1
discord==2.3.2
discord.py==2.3.2
distro==1.8.0
elevenlabs==0.2.26
exceptiongroup==1.2.0
executing==2.0.1
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
multidict==6.0.4
openai==1.3.7
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.5.2
pydantic-core==2.14.5
pygments==2.17.2
PyNaCl==1.5.0
python-dotenv==0.21.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
tqdm==4.66.1
traitlets==5.14.0
typing-extensions==4.8.0
urllib3==2.1.0
wcwidth==0.2.12
websockets==12.0
yarl==1.9.4
